requirements.txt
langchain-core==1.1.1
langchain-community==0.4.1
langchain-openai==1.1.0
langgraph==1.0.4
python-dotenv==1.2.1
requests==2.32.5
beautifulsoup4==4.14.3
aiohttp==3.13.2
typing-extensions==4.15.0
pydantic==2.12.5
google-genai==1.53.0